attempt at dynamic imports failed... can't load source map file.
This commit is no good... but I wanted to show what was attempted. I ultimately wanted to lazy load locale options but it keeps wanting to grab the source map file and doesn't know how to load it. It's also more complicated than what I started with but I was just trying different options.