ARK-4 post card (#2)
* init post card

* match master

* separate variants into own files

* pass through types

* lint and 100% test coverage

* naming things

* attempt at dynamic imports failed... can't load source map file.

This commit is no good... but I wanted to show what was attempted. I ultimately wanted to lazy load locale options but it keeps wanting to grab the source map file and doesn't know how to load it. It's also more complicated than what I started with but I was just trying different options.

* time ago loads, but not particularly efficiently

* 100 percent coverage

* hopefully this compiles

* maybe this compiles with ci

* address items in review

* Update time-ago.docs.mdx

* cleanup

* pnpm update

* make buttonText in hero component optional

* rearrange things. change how types are re-exported

* postcard

* date fns

* lint and pnpm

Co-authored-by: Tataihono Nikora <tataihono.nikora@gmail.com>